add, edit, listのテストをTaskUIControllerに集約しました。
